raw conversion of FixtureInitialization specs
